Refactor clang runtime bits support for zig integration
